+++ scripts-src/pre.sh
@@ -2,9 +2,10 @@
 
 KVER="${KVER:-$(uname -r)}"
 
-if [ -e "/usr/lib/modules/${KVER}" ]; then
+if [ -e /usr/lib/modules/"$KVER" ] && [ ! -e /usr/lib/modules/backup/"$KVER" ]; then
+	[ -d /usr/lib/modules/backup ] || mkdir /usr/lib/modules/backup
 	cp -a \
-		"/usr/lib/modules/${KVER}" \
-		/usr/lib/modules/backup/ \
+		/usr/lib/modules/"$KVER" \
+		/usr/lib/modules/backup/"$KVER" \
 	;
 fi
